added layer manager actions. Moved layer services to local folder instead of core
