+++ src/app/packages/core/core.module.ts
@@ -1,10 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { DataLayersService } from './services/data-layers.service';
 
 @NgModule({
-  imports: [CommonModule],
-  providers: [DataLayersService]
+  imports: [CommonModule]
 })
 
 export class CoreModule { }
