+++ src/app/packages/core/index.ts
@@ -1,6 +1,4 @@
 export { CoreModule } from "./core.module";
 export { MenuItem } from "./models/menu-item.model";
-export { DataLayersService } from './services/data-layers.service';
-export { ILayerTreeNode } from './models/data-layers/layer-tree-node';
 export { MenuActionTypes, AddMenuItemAction, SelectMenuItemAction, UnSelectMenuItemAction, AnimationStartAction, AnimationEndAction } from './actions/menu.actions';
 export { Position } from './models/position.model';
