+++ src/app/packages/menu-items/layers-manager/components/layer-tree/layer-tree.component.scss
