+++ src/app/packages/menu-items/layers-manager/components/layer-tree/layer-tree.component.spec.ts
@@ -2,9 +2,9 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { DebugElement } from '@angular/core';
 import { By } from '@angular/platform-browser';
 import { LayerTreeComponent } from './layer-tree.component';
-import { LayersManagerModule } from '../layers-manager.module';
+import { LayersManagerModule } from '../../layers-manager.module';
 import { CoreModule } from '@ansyn/core/core.module';
-import { ILayerTreeNode } from '@ansyn/core';
+import { ILayerTreeNode } from '../../models/layer-tree-node';
 import { TreeNode } from 'angular-tree-component';
 import { Observable } from 'rxjs/Observable';
 import { StoreModule } from '@ngrx/store';
