+++ src/app/packages/menu-items/layers-manager/components/layers-manager/layers-manager.component.scss
